+++ nichols-works/metadata.rb
@@ -1,0 +1,15 @@
+name "nichols-works"
+maintainer "Daniel Nichols"
+maintainer_email "dan@nichols.works"
+license "mit"
+description "Does what I need it to do"
+long_description "Does what I need it to do"
+version "0.2.0"
+
+depends "apt"
+depends "chef-apt-docker"
+depends "docker"
+depends "cloudcli"
+depends "tar"
+
+supports "ubuntu"